修改repo算子精度

+++ src/ops/rope/cpu/rope_cpu.cpp
@@ -25,8 +25,8 @@ void rope_impl_(T *out,
             const size_t head_base = pos_base + h * head_dim;
 
             for (size_t j = 0; j < half_dim; ++j) {
-                const double exponent = (2.0 * static_cast<double>(j)) / d;
-                const double inv_freq = std::pow(static_cast<double>(theta), -exponent);
+                const double exponent = (2.0 * static_cast<float>(j)) / d;
+                const double inv_freq = std::pow(static_cast<float>(theta), -exponent);
                 const double phi = p * inv_freq;
                 const float s = static_cast<float>(std::sin(phi));
                 const float c = static_cast<float>(std::cos(phi));
